feat(create-price-feed): Add InsuredBridge (#3388)
* feat(create-price-feed): Add InsuredBridge

* Should return YES for expired relays one cannot dispute

* Update CreatePriceFeed.ts

* Update CreatePriceFeed.js

* Fix script

* Fix CreatePriceFeed test

* fix

* Update Relayer.ts

* fix test

* Update 1.json

* Update CreatePriceFeed.js

* Fix script

* Update InsuredBridgePriceFeed.ts

* Update CreatePriceFeed.js

* Update InsuredBridgePriceFeed.js

* fix

* simplify test

+++ packages/core/test/insured-bridge/BridgePool.js
@@ -134,7 +134,7 @@ describe("BridgePool", () => {
   // Generate ABI encoded deposit data and deposit data hash.
   const generateDepositHash = (_depositData) => {
     const depositDataAbiEncoded = web3.eth.abi.encodeParameters(
-      ["uint8", "uint64", "address", "address", "address", "uint256", "uint64", "uint64", "uint64"],
+      ["uint256", "uint64", "address", "address", "address", "uint256", "uint64", "uint64", "uint64"],
       [
         _depositData.chainId,
         _depositData.depositId,
@@ -181,7 +181,7 @@ describe("BridgePool", () => {
   // Replicate the hashed ancillary data that is returned by BridgePool's internal _getRelayHash() method.
   const generateRelayAncillaryDataHash = (_depositData, _relayData) => {
     const parameters = [
-      { t: "uint8", v: _depositData.chainId },
+      { t: "uint256", v: _depositData.chainId },
       { t: "uint64", v: _depositData.depositId },
       { t: "address", v: _depositData.l1Recipient },
       { t: "address", v: _depositData.l2Sender },
